Add Legesher to the list
From https://github.com/legesher/legesher/issues/54
With the Legesher project, there are a lot of opportunities for people
of all backgrounds to contribute the skills they either want to grow in
or give back. With a strong emphasis on written language translation,
individuals who are bilingual have a great space to contribute.
Individuals who want to grow in community development, marketing,
writing impactful pieces of insight, or just want to learn more about
the tech world have ample amounts of opportunities here as well!